Avoid redefining components during rendering
This change avoids the definition of new components in `render()` calls,
but instead moved them out into their own components or memoises the
definitions.

If we didn't do that, React would recreate the state of each of these
components on every call, which is a performance issue and could lead to
inconsistencies if the components have a state that should be preserved.